+++ px7_music/main.py
@@ -3,7 +3,7 @@ import px7_music.core.handler               as Handler
 import px7_music.player.auto_play_mode      as AP
 import px7_music.player.playback            as Playback
 
-from px7_music.config           import ERROR_TRACBACK
+from px7_music.config           import ERROR_TRACEBACK
 from px7_music.core             import latency
 from px7_music.core.parser      import CommandParser
 from px7_music.player.player    import get_player
@@ -11,7 +11,7 @@ from px7_music.utility.docs     import get_installation_guide, get_help_text
 from px7_music.utility.utils    import ANSI, Preloader, clear_screen
 
 
-sys.tracebacklimit = ERROR_TRACBACK
+sys.tracebacklimit = ERROR_TRACEBACK
 cmd_parser  =   CommandParser()
 spinner     =   Preloader()
 
@@ -23,6 +23,8 @@ def register_commands():
     cmd_parser.register("search",   Handler.search_handler) # search and fills the queue {supports flag}
     cmd_parser.register("/s",       Handler.search_handler) # search and fills the queue {supports flag}
     cmd_parser.register("play",     Handler.play_handler)   # play <index from queue>
+    cmd_parser.register("fav",      Handler.fav_handler)
+    cmd_parser.register("favs",     Handler.favs_handler)
     cmd_parser.register("exit",     Handler.exit_handler)   # exits the program
 
     cmd_parser.register("current",  Playback.show_current)  # shows info of current playing track
@@ -32,8 +34,10 @@ def register_commands():
     cmd_parser.register("queue",    Playback.show_queue)    # shows current queue
     cmd_parser.register("resume",   Playback.resume)        # resume track
     cmd_parser.register("pause",    Playback.pause)         # pause track
+    cmd_parser.register("load",     Playback.load)          # loads last searched result into queue
+    cmd_parser.register("shuffle",  Playback.shuffle_queue) # shuffles the queue respecting the current playing track
 
-    cmd_parser.register("help",     get_help_text)
+    cmd_parser.register("help",     get_help_text)          # detailed documentation of available commands
 
     cmd_parser.register("latency",  check_network)          # shows network latency
     cmd_parser.register("clear",    clear_screen)           # clears the terminal and prints banner